+++ src/reducers/initialState.js
@@ -2,10 +2,12 @@ export default {
     signUp: {
         loading: false,
         success: false,
+        rejected: false
       },
     logIn: {
         loading: false,
         success: false,
-        rejected: false
+        rejected: false,
+        token:''
     }
 }
