+++ src/reducers/rootReducer.js
@@ -1,5 +1,6 @@
 import {combineReducers} from 'redux';
-import {signUp,logIn} from './trial'
+import {signUp} from './signup.reducer';
+import {logIn} from './login.reducer';
 
 export default combineReducers({
 signUp,
